cpp: Put DeviceAllocationTypeSet in the cheapest possible header
