+++ cpp/src/arrow/array/array_test.cc
@@ -45,6 +45,7 @@
 #include "arrow/buffer.h"
 #include "arrow/buffer_builder.h"
 #include "arrow/compare.h"
+#include "arrow/device_allocation_type.h"
 #include "arrow/result.h"
 #include "arrow/scalar.h"
 #include "arrow/status.h"
