+++ cpp/src/arrow/buffer_test.cc
@@ -29,6 +29,7 @@
 #include "arrow/buffer.h"
 #include "arrow/buffer_builder.h"
 #include "arrow/device.h"
+#include "arrow/device_allocation_type.h"
 #include "arrow/io/interfaces.h"
 #include "arrow/memory_pool.h"
 #include "arrow/status.h"
